build: Default Snakemake's --cores option to "all"
Applies when our own --cpus option isn't provided.  This will allow
upgrading of Snakemake in our Docker runtime without inflicting the
addition of --cores (or equivalent) arguments onto every existing
`nextstrain build` invocation.

Snakemake requires the --cores option as of 5.11.¹  This has spawned
much discussion² where it was made clear this is an intentional,
permanent change and a default will not be added.  Add our own default
instead to insulate our users from this and make Nextstrain builds
fast-by-default.  Our --cpus option can be used to limit CPU usage back
from this default if necessary, and users can always specify --cores (or
equivalents) themselves.

Note that we don't default the --cpus option (opts.cpus) because we
don't want to _limit_ the CPUs available in the Docker or AWS Batch
runtimes, just inform Snakemake that it can use whatever CPUs are
available.³  This means that `nextstrain build` will parallelize as much
as it can regardless of the runner, which feels right in the context of
launching a multi-step workflow.  Note that this is different than
Augur's default of using a single CPU, but the context in which Augur is
invoked is typically a single-step in a larger workflow where more
explicit coordination between steps is required for parallelism.

¹ https://github.com/snakemake/snakemake/issues/283

² https://github.com/snakemake/snakemake/issues/308
  https://github.com/snakemake/snakemake/issues/312
  https://github.com/snakemake/snakemake/issues/450
  https://github.com/snakemake/snakemake/issues/885

³ But there's a caveat here for AWS Batch, see
  <https://github.com/nextstrain/cli/issues/175>.